% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sigProb.r
\name{plot.sigProb}
\alias{plot.sigProb}
\title{Plot predicted probabilities from a \code{sigProb} object.}
\usage{
\method{plot}{sigProb}(object, prob, xvar, col = "blue", ylab, xlab,
  pch = 16, main = "", ...)
}
\arguments{
\item{object}{an object of class \code{sigProb}, which is obtained by using \code{\link{predict.sigfit}} on a model fit using \code{\link{sigint}}.}

\item{prob}{A string providing the column name for the column of \code{object$predicted} should be used as the outcome or probability of interest.}

\item{xvar}{A string providing the column name of the column (from either object$model or object$par) that provides the "x-variable" in the plot.}

\item{col}{A string used to}
}
\description{
This method takes a \code{sigProb} object produced by
\code{\link{predict.sigfit}} and plots the comparative static(s) of interest.
}
\examples{
data(sanctionsData)

f1 <- sq+cd+sf+bd ~ sqrt(senderecondep) + senderdemocracy + contig + ally -1|#SA
                    anticipatedsendercosts|#VA
                    sqrt(targetecondep) + anticipatedtargetcosts + contig + ally|#CB
                    sqrt(senderecondep) + senderdemocracy + lncaprat | #barWA
                    targetdemocracy + lncaprat| #barWB
                    senderdemocracy| #bara
                    -1#VB
## Using Nested-Pseudo Likelihood  with default first stage                 
fit1 <- sigint(f1, data=sanctionsData, npl.trace=TRUE)

new.theta <- data.frame(t(replicate(25, coef(fit1))))
new.theta[,19] <- seq(-6, 0, length=25)

p1.a <- predict(fit1, newdata=sanctionsData[93,], new.theta = new.theta,
                control=list(gridsize=500))

plot(p1.a, prob="pc", ylab="Pr Challenge", xlab="Audience Costs", type="l")



}
\seealso{
\code{\link{predict.sigfit}} \code{\link{generate.eq}}
}
