% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sigfit.r
\name{print.summary.sigfit}
\alias{print.summary.sigfit}
\title{Print the summary table for a \code{sigfit} object.}
\usage{
\method{print}{summary.sigfit}(x, ...)
}
\arguments{
\item{x}{a \code{summary.sigfit} object.}

\item{...}{Additional arguments (not currently used)}
}
\description{
Prints the summary regression table for a model fitted with \code{\link{sigint}}.
}
\details{
Prints the standard regression results table from a fitted strategic model,
along with the log-likelihood and number of games used in estimation.
}
