\name{sigInt-package}
\alias{sigInt-package}
\alias{sigInt}
\docType{package}
\title{
\packageTitle{sigInt}
}
\description{
\packageDescription{sigInt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sigInt}
\packageIndices{sigInt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sigInt}

Maintainer: \packageMaintainer{sigInt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
