% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.r
\docType{data}
\name{SIGMA}
\alias{SIGMA}
\title{First-stage covariance matrix}
\usage{
data(SIGMA)
}
\description{
Bootstrapped covariance matrix (\eqn{836 \times 836}) for
 \code{sanctionsData$PRhat} and \code{sanctionsData$PFhat}.
 Rows and columns are ordered \eqn{[p_{R,d}, p_{F,d}]}, for 
 games \eqn{d=1,...,418}.
Only used for examples and replication.
}
\seealso{
\code{\link{sanctionsData}}, \code{\link{sigint}}
}
\keyword{data}
