% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sigfit.r
\name{summary.sigfit}
\alias{summary.sigfit}
\title{Summarize a fitted crisis signaling model}
\usage{
\method{summary}{sigfit}(object, vcov, ...)
}
\arguments{
\item{object}{a fitted model of class \code{sigfit}}

\item{vcov}{a substitute variance covariance matrix}

\item{...}{Additional arguments (not currently used)}
}
\value{
An object of class \code{summary.sigfit}.  This object
contains the information needed to print the summary
}
\description{
The default method for summarizing a \code{sigfit} object.
}
\details{
Forms a block regression results table from fitted crisis signaling model.
}
\seealso{
\code{\link{print.summary.sigfit}}.
}
